requirements.txt
asgiref==3.8.1
blinker==1.6.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
filelock==3.10.4
Flask==2.3.2
idna==3.7
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
platformdirs==3.1.1
psycopg2-binary==2.9.9
pyparsing==3.0.9
pyswip==0.2.9
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.21.0
Werkzeug==2.3.4
